Update README.md
Ace Autos Car Rentals API is a backend service built with Node.js and Express that enables seamless management of car listings, user accounts, and booking transactions for a car rental business. 

It integrates with Paystack for secure payment processing and sends automated email notifications to customers upon successful or failed payments. This API is designed to provide a robust foundation for a car rental platform with essential features such as user authentication, booking management, and real-time payment status updates.